backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
bcrypt
jwt
faiss-cpu
sentence-transformers
openai